Added example.html, fixed crash based on recent type change.

+++ BranchSDK-Samples/Windows/TestBed/TextField.cpp
@@ -73,7 +73,7 @@ TextField::appendText(const std::wstring& text, size_t maxLength)
 	wstring newText = getText() + L"\r\n" + text;
 	// Limit text length
 	const wstring::size_type total = newText.length();
-	const size_t offset = total - maxLength;
+	const __int64 offset = total - maxLength;
 	if (offset > 0)
 	{
 		newText = newText.substr(offset, maxLength - 1);
